Extract useffects into DataContext
